LocationService/requirements.txt
# LocationService/requirements.txt
fastapi
uvicorn[standard]
redis
python-dotenv
azure-identity